🧬 Migrate the data models about settings in section *<API>.http.request* to module *api_config.apis.request*.
