+++ pymock_api/model/swagger_config.py
@@ -5,10 +5,9 @@ from abc import ABCMeta, abstractmethod
 from pydoc import locate
 from typing import Any, Dict, List, Optional, Union
 
-from pymock_api.model import APIConfig
-from pymock_api.model import APIParameter as PyMockAPIParameter
-from pymock_api.model import MockAPI, MockAPIs
+from pymock_api.model import APIConfig, MockAPI, MockAPIs
 from pymock_api.model.api_config import BaseConfig, _Config
+from pymock_api.model.api_config.apis import APIParameter as PyMockAPIParameter
 from pymock_api.model.enums import ResponseStrategy
 
 Self = Any
