+++ pymock_api/server/application/process.py
@@ -3,7 +3,8 @@ from abc import ABC, ABCMeta, abstractmethod
 from pydoc import locate
 from typing import Any, Dict, List, Union, cast
 
-from ...model import APIParameter, HTTPRequest, HTTPResponse, MockAPI
+from ...model import HTTPResponse, MockAPI
+from ...model.api_config.apis import APIParameter, HTTPRequest
 from ...model.enums import ResponseStrategy
 from .request import BaseCurrentRequest
 from .response import BaseResponse
