+++ test/integration_test/server/application/init.py
@@ -9,7 +9,8 @@ from fastapi.testclient import TestClient as FastAPITestClient
 from flask.app import Response as FlaskResponse
 from httpx import Response as FastAPIResponse
 
-from pymock_api.model import APIConfig, APIParameter, MockAPI, load_config
+from pymock_api.model import APIConfig, MockAPI, load_config
+from pymock_api.model.api_config.apis import APIParameter
 from pymock_api.server.application import BaseAppServer, FastAPIServer, FlaskServer
 from pymock_api.server.application.response import HTTPResponse as _HTTPResponse
